Create an aggregation vector

maturing

Usage

agg_vec(x = character(), aggregated = logical(vec_size(x)))

Arguments

x

The vector of values.

aggregated

A logical vector to identify which values are <aggregated>.

Details

An aggregation vector extends usual vectors by adding <aggregated> values. These vectors are typically produced via the aggregate_key() function, however it can be useful to create them manually to produce more complicated hierarchies (such as unbalanced hierarchies).

Examples

agg_vec(
  x = c(NA, "A", "B"),
  aggregated = c(TRUE, FALSE, FALSE)
)
#> <agg_vec[3]>
#> [1] <aggregated> A            B